renamer provider og søknadtype for overgangsstønad til å bli mer presis
