+++ src/context/ContextProviders.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { PersonProvider } from './PersonContext';
-import { SøknadProvider } from './SøknadContext';
+import { OvergangsstønadSøknadProvider } from '../overgangsstønad/OvergangsstønadContext';
 import { TogglesProvider } from './TogglesContext';
 import { BarnetilsynSøknadProvider } from '../barnetilsyn/BarnetilsynContext';
 import { SkolepengerSøknadProvider } from '../skolepenger/SkolepengerContext';
@@ -14,11 +14,11 @@ const ContextProviders: React.FC<{ children?: React.ReactNode }> = ({
     <TogglesProvider>
       <GjenbrukProvider>
         <PersonProvider>
-          <SøknadProvider>
+          <OvergangsstønadSøknadProvider>
             <BarnetilsynSøknadProvider>
               <SkolepengerSøknadProvider>{children}</SkolepengerSøknadProvider>
             </BarnetilsynSøknadProvider>
-          </SøknadProvider>
+          </OvergangsstønadSøknadProvider>
         </PersonProvider>
       </GjenbrukProvider>
     </TogglesProvider>
