+++ src/overgangsstønad/steg/2-bosituasjon/Bosituasjon.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import { useLokalIntlContext } from '../../../context/LokalIntlContext';
-import { useSøknad } from '../../../context/SøknadContext';
+import { useOvergangsstønadSøknad } from '../../OvergangsstønadContext';
 import { IBosituasjon } from '../../../models/steg/bosituasjon';
 import { useLocation } from 'react-router-dom';
 import { erFerdigUtfylt } from '../../../helpers/steg/bosituasjon';
@@ -9,7 +9,7 @@ import { RoutesOvergangsstonad } from '../../routing/routesOvergangsstonad';
 import { hentPathOvergangsstønadOppsummering } from '../../utils';
 import Side, { ESide } from '../../../components/side/Side';
 import { Stønadstype } from '../../../models/søknad/stønadstyper';
-import { ISøknad } from '../../../models/søknad/søknad';
+import { SøknadOvergangsstønad } from '../../../models/søknad/søknad';
 import { logSidevisningOvergangsstonad } from '../../../utils/amplitude';
 import { useMount } from '../../../utils/hooks';
 import { kommerFraOppsummeringen } from '../../../utils/locationState';
@@ -21,7 +21,7 @@ const Bosituasjon: FC = () => {
     settSøknad,
     settDokumentasjonsbehov,
     mellomlagreOvergangsstønad,
-  } = useSøknad();
+  } = useOvergangsstønadSøknad();
   const bosituasjon = søknad.bosituasjon;
   const location = useLocation();
   const kommerFraOppsummering = kommerFraOppsummeringen(location.state);
@@ -32,7 +32,7 @@ const Bosituasjon: FC = () => {
   useMount(() => logSidevisningOvergangsstonad('Bosituasjon'));
 
   const settBosituasjon = (bosituasjon: IBosituasjon) => {
-    settSøknad((prevSoknad: ISøknad) => {
+    settSøknad((prevSoknad: SøknadOvergangsstønad) => {
       return {
         ...prevSoknad,
         bosituasjon: bosituasjon,
